daytistics-core/pyproject.toml
[tool.poetry]
name = "daytistics-core"
version = "0.1.0"
description = ""
authors = ["Leo Gall <gall.dev@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-allauth = "^64.0.0"
Django = "^5.0.7"
django-widget-tweaks = "^1.5.0"
django-mathfilters = "^1.0.0"
whitenoise = { extras = ["brotli"], version = "^6.7.0" }
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.4.0"
celery = "^5.4.0"
django-ninja-jwt = "^5.3.4"
django-ninja = "^1.3.0"
mypy = "^1.11.2"
django-stubs = "^5.1.0"
pytz = "^2024.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.7.0"
faker = "^26.1.0"
pytest-env = "^1.1.3"
freezegun = "^1.5.1"
pylint = "^3.2.6"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
adopts = ["-vv", "-rP"]

[tool.pytest_env]
DJANGO_SECRET_KEY = "RtcPToOOZ82erjag0zAVAQ3mmha9MrTx"
DJANGO_ALLOWED_HOSTS = "localhost,example.com"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
